fix(event): use correct types for window proc params

+++ src/event/linux/Window.cpp
@@ -1,5 +1,5 @@
 #include "event/Window.hpp"
 
-int32_t OsWindowProc(void* window, uint32_t message, uint32_t wparam, int32_t lparam) {
+int32_t OsWindowProc(void* window, uint32_t message, uintptr_t wparam, intptr_t lparam) {
     return 0;
 }
